refactor: move override tests to standalone file.
